requirements.txt
dominate==2.9.0
einops==0.7.0
imageio==2.33.0
lpips==0.1.4
matplotlib==3.5.2
numpy==1.22.3
opencv_python==4.8
Pillow==9.2.0
PyYAML==6.0
scipy==1.3.1
tensorboardX==2.2
tensorflow==2.9.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.66.1
visdom==0.2.4
scikit-image==0.17.2
